OddDataUnits to float

+++ utils/structs.go
@@ -193,7 +193,7 @@ type PickStruct struct {
 }
 type PickDetails struct {
 	Picks           []Picks   `json:"picks"`
-	OddDataUnits    int       `json:"oddDataUnits"`
+	OddDataUnits    float64       `json:"oddDataUnits"`
 	OddDataWins     int       `json:"oddDataWins"`
 	PickId          string    `json:"pickId"`
 	Gamestatus      []string  `json:"gamestatus"`
@@ -206,7 +206,7 @@ type Picks struct {
 	TeamName         string `json:"teamName"`
 	OppTeamName      string `json:"oppTeamName"`
 	Sport            string `json:"sport"`
-	OddDataMoneyLine float64    `json:"oddDataMoneyLine"`
+	OddDataMoneyLine int    `json:"oddDataMoneyLine"`
 }
 type AwsPick struct {
 	GameStatus      bool          `json:"gamestatus.NULL"`
